feat(bin): check yt-dlp executable exists
Before doing any work, we're checking for the existence of yt-dlp.
The logic for checking was moved to a separate module -- note most of
the feature was copied from the external lib (see comments in
lib/check-executable); the reason for this: I'd prefer not to depend on
the external lib .. the logic is fairly simple, and we're using the sync
implementation only.