fix(anime-dl): avoid recursive target directory
We were creating a 'recursive' download point by joining the passed-in
directory name & anime dir name.  On Mac, this was throwing an error &
aborting the program.  The weird thing is, on Linux, this was resolved
correctly -- this is unexpected & should be tested (perhaps this is a
bug).

Fix by removing join method.