fix(anime-dl): episode filename generation
Previously, a loop index was being using for filename generation, instead
of using the actual, currently processed episode's number.

This commit also removes the intermediate step of creating the series
episodeUrl list in memory.  For larger series, this is really bad on
memory.  Instead, since we have the current episode number we want to
process, we're generating the episodeURL at the current iteration
episode range.