fix(players): premature player destroy
When not in debug mode & processing the child process' stdout
the player status logger was timing out & subsequently killing
the player instance (BAD!).  This refactors the status logger out, and just
writes to the main process's stdout (with in-place returns, to avoid
muddying up the output).  We are also removing the status logger for now
-- refactors are in the works.